dbx-project/verify_keys.py

import os
from dagster._core.pipes.context import PipesSession
from dagster_databricks.pipes import PipesUnityCatalogVolumesContextInjector, PipesUnityCatalogVolumesMessageWriter
from databricks.sdk import WorkspaceClient

# Mock WorkspaceClient since we just need the class structure, or we can just mock dependencies
# Actually, we can just inspect what keys generated if we can instantiate it.
# We need a dummy client.

class DummyClient:
    pass

try:
    injector = PipesUnityCatalogVolumesContextInjector(client=DummyClient(), volume_path="/Volumes/x/y/z")
    writer = PipesUnityCatalogVolumesMessageWriter(client=DummyClient(), volume_path="/Volumes/x/y/z")
    
    # We can't easily open a full session without mocking context data.
    # But we can check if we can simulate the params.
    
    # pipes_session.get_bootstrap_env_vars() calls params_loader.
    # By default PipesSession uses PipesEnvVarParamsLoader.
    
    # Let's just peer into how open_pipes_session works or documentation.
    # Or just use the fact that I can run this script with dummy data.
    
    pass
except Exception as e:
    print(e)
    
# Alternative: Check dagster source code for PipesSession.get_bootstrap_env_vars
from dagster._core.pipes.utils import _make_bootstrap_env_vars
# This function is internal.

# Let's try to verify by inspecting the module code via cli if possible, or just reading source.
